Simplify the sum in the asymptotic version of T01
The previous version was what came out directly from Mathematica. I
realised that it didn't make use of m being an integer and in that
case it's just a polynomial which we can handle trivially. It's not a
strictly necessary optimization in this case but it will probably help
with bounding the tail.