Messy commit with formatting and unfinished changes
I had several old changes that I started with before summer that were
not finished. Some of these were supposed to be finished and some of
them was of a more temporary nature. Then I ran the formatter and the
changes from that got mixed up with the other changes. In addition to
this I also changed all prec to precision since the former has been
deprecated.

Instead of figuring out exactly what should be commited and what
should be discarded I make this commit now and will have to clear up
afterwards.