Improve computation of asymptotic defect for BH
Tune the starting depth for on of the intervals. Improve the
performance of F0 by reducing the number of allocations and reusing
computations. This reduces the number of allocations of F0 by about
one third, some more improvements would be possible but it would
likely be more beneficial to improve the methods in ArbExtras instead.