Update weight for BHKdV
The new weight uses ℯ inside the log instead of 10. The new weight
gives a slightly larger value for C_B but a much lower alpha_0,
resulting in an improved value in the end. The choice of precisely ℯ
mostly for it to look nice, 2 or 3 would likely have worked just as
well.