Add first lemmas and use them in alpha0 for BHKdV
Lemmas as methods which don't do anything except reference certain
properties that needs to be used in different parts of the code. The
are somewhat of an experiment for now but their main goal is to make
it to simplify handling of statement that occurs in several methods,
giving one place to refer to.