Update integration for BHKdV to use better enclosure of alpha
Before the enclosure was computed with Arb((-1, -1 + epsilon)), this
in general makes it so that the lower endpoint is not exactly -1. By
rewriting it so that the lower bound is exactly -1 we get much better
enclosures. In particular this makes it so that we can remove the
assumption on monotonicity in the parameter for the integrands in T012
and T022. It does take longer to compute in this case though, so there
are drawbacks.