Make all calls to factorial use Int instead of Arb
This is done in preparation for updating SpecialFunctions to version
2.0 where it no longer aliases factorial(n) to gamma(n + 1). While we
could define factorial(n::Arb) = gamma(n + 1) ourselves it should be
safe to use the factorial with Int instead, if the argument is to
large it throws an error so it will never give an incorrect result.
The only risk would be if the result was used to promote something
else to Arb, but that doesn't seem to be the case anywhere.