Clausen: Improve clausenc(x, s, β)
It now uses _reduce_argument_clausen and handles wide values of x
better for s not equal to 0, 2 or 3.

For s = 0, 2 or 3 and wide values of s it can now be slightly worse.
But we no longer have to prove that there is a unique critical point.

Also add more and improved tests for clausenc in general.