KdV: Rework find_good_as and the way it is used in the construction
It now takes a step-range of N0 values to consider, instead of just an
upper bound. This makes it possible to start at a higher value and
take larger steps. It also adds an argument for returning the defects.

To work with this the default parameter choice in the construction is
given as a step-range instead of just an upper bound.