Clausen: Temporary commit - improve enclosure for s::ArbSeries
This is code written when trying to improve the compute enclosure for
BHKdV. It improves the enclosure for clausenc and clausens with
s::ArbSeries close to an integer and wide x. In the end it seems like
it is not needed and will therefore be reversed. But if it ever is
needed in the future this commit is here.