KdV: Simplify _F0_bhkdv, removing some prior optimizations
It turns out that the implementation of _F0_bhkdv was not what was
limiting the performance. The issue was instead
ArbExtras.enclosure_polynomial. When the input polynomial was wide it
took a lot of work for it to try, but fail, to isolate the roots of
the derivative. The performance issues will therefore be solved by
updating ArbExtras.enclosure_series instead.