KdV: Pick parameters based on midpoint of ball
Previously the parameters were picked based on the full ball. This
meant that the two balls coming from the nearby endpoints of two
subintervals would in general get the same parameters, since they
overlap. By picking it based on the midpoint this is no longer the
case. This gives a more consistent behaviour for each subinterval.