Clausen: Restructure _clausenc_zeta and _clausens_zeta
The logic for handling the removable singularities has been factored
out into separate methods. The main benefit with doing this is that we
can now use ArbExtras.enclosure_series for the full function in all
cases and in general it only uses the logic for handling the removable
singularities when computing the remainder term. This gives better
enclosures for wide balls near the removable singularities.