KdV: Improve D0_bounded_by by checking monotonicity of u0 once
Close to -1 it is the evaluation of u0 and not of the integral which
is the most time consuming, it is therefore worth it to optimize that
a bit more.

In practice u0 is increasing on the interval and before the evaluation
was done using ArbExtras.enclosure_series which in most cases picked
up this monotonicity. But it meant having to compute the derivative a
lot of times. Now we instead check that it is monotone on (most of)
the interval from the start and then we don't have to compute the
derivative as many times.