BHKdV: Remove testing code for developing bound of G2
This code was used when developing that bounds are not in the paper.
It was very helpful for finding typos and figuring out what seemed to
work and what would probably not work. It is however not part of the
computer assisted proof and is therefore removed. Could still be
helpful to look back at though.